flask_deepfake_app/requirements.txt
Flask
torch
torchvision
opencv-python-headless  
Pillow
Werkzeug              